did some stuff, many haven't finished

+++ c/arithmetic-op/Problem AC/solution.c
@@ -1,16 +1,16 @@
 #include <stdio.h>
 
 int main(){
-    int teams, chances=1, coin=2; 
+    long long int teams, chances=1, coin=2; 
 
-    scanf("%d", &teams); 
+    scanf("%lld", &teams); 
     
     while (teams != 0){
         chances *= coin; 
         teams--;
     }
     chances--;
-    printf("%d\n", chances);
+    printf("%lld\n", chances);
 
     return 0; 
 }
